+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## v3.6.3
  * improve javascript
+ * fix problem when a token error occurs
 
 ## v3.6.2
  * add detectCardBrand